+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- `drawer` 组件 `with-header` 配置项改为 `showHeader`
+- `drawer` 组件新增 `showCloseBtn` 和 `showFooter` 配置项
+
 ## 0.7.0-alpha.4 (2022-10-11)
 
 - 新增 `checkbox` 组件